Extracted a JSON response method and Standarized Error method
