+++ hub/web_app/server/main.go
@@ -1,6 +1,5 @@
 package main
 
-
 import (
 	"fmt"
 	"log"
