--- hub/web_app/server/models/claims.go
@@ -1,8 +1,0 @@
-package models
-
-import "github.com/golang-jwt/jwt/v4"
-
-type Claims struct{
-	Username string `json:"username"`
-	jwt.RegisteredClaims
-}
