+++ hub/web_app/server/models/models.go
@@ -1,0 +1,18 @@
+package models
+
+import "github.com/golang-jwt/jwt/v4"
+
+type Claims struct{
+	Username string `json:"username"`
+	jwt.RegisteredClaims
+}
+
+//simulated database
+var ValidUsers = map[string]string{
+	"testuser":"password123",//username: password
+}
+
+type Credentials struct{
+	Username string `json:"username"`
+	Password string `json:"password"`
+}
