+++ hub/web_app/server/utils/jwt.go
@@ -29,3 +29,4 @@ func GenerateJWT(username string) (string,error){
 	return token.SignedString(Jwtsecret)
 
 }
+
