+++ hub/web_app/server/utils/response.go
@@ -1,0 +1,16 @@
+package utils
+
+import (
+	"encoding/json"
+	"net/http"
+)
+
+func RespondJSON(w http.ResponseWriter,data interface{}){
+	w.Header().Set("Content-Type","application/json")
+	json.NewEncoder(w).Encode(data)
+}
+
+func RespondError(w http.ResponseWriter,message string, statuscode int){
+	w.WriteHeader(statuscode)
+	RespondJSON(w,map[string]string{"error":message})
+}
